+++ Source/Scene/Cesium3DTilesetMetadata.js
@@ -16,8 +16,7 @@ import TilesetMetadata from "./TilesetMetadata.js";
  * </p>
  *
  * @param {Object} options Object with the following properties:
- * @param {Object} [options.tilesetJson] The tileset JSON object. If this is undefined, then options.extension must be defined.
- * @param {Object} [options.extension] The JSON object, used for backwards compatibility with the 3DTILES_metadata extension.
+ * @param {Object} options.tilesetJson The tileset JSON object.
  * @param {MetadataSchema} options.schema The parsed schema.
  *
  * @exception {DeveloperError} One of tilesetJson and extension must be defined.
@@ -30,27 +29,18 @@ import TilesetMetadata from "./TilesetMetadata.js";
 function Cesium3DTilesetMetadata(options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
   const tilesetJson = options.tilesetJson;
-  const extension = options.extension;
 
   // The calling code is responsible for loading the schema.
   // This keeps metadata parsing synchronous.
   const schema = options.schema;
 
   //>>includeStart('debug', pragmas.debug);
-  if (defined(tilesetJson) === defined(extension)) {
-    throw new DeveloperError(
-      "One of options.tilesetJson and extension must be defined."
-    );
-  }
+  Check.typeOf.object("options.tilesetJson", tilesetJson);
   Check.typeOf.object("options.schema", schema);
   //>>includeEnd('debug');
 
-  let metadata;
-  if (defined(tilesetJson)) {
-    metadata = tilesetJson.metadata;
-  } else {
-    metadata = extension.tileset;
-  }
+  // An older schema stored the tileset metadata in the "tileset" property.
+  const metadata = defaultValue(tilesetJson.metadata, tilesetJson.tileset);
 
   let tileset;
   if (defined(metadata)) {
@@ -60,9 +50,7 @@ function Cesium3DTilesetMetadata(options) {
     });
   }
 
-  const definedJson = defaultValue(tilesetJson, extension);
-  const groupsJson = definedJson.groups;
-
+  const groupsJson = tilesetJson.groups;
   const groups = {};
   if (defined(groupsJson)) {
     for (const groupId in groupsJson) {
@@ -81,9 +69,9 @@ function Cesium3DTilesetMetadata(options) {
   this._groups = groups;
   this._tileset = tileset;
 
-  this._statistics = definedJson.statistics;
-  this._extras = definedJson.extras;
-  this._extensions = definedJson.extensions;
+  this._statistics = tilesetJson.statistics;
+  this._extras = tilesetJson.extras;
+  this._extensions = tilesetJson.extensions;
 }
 
 Object.defineProperties(Cesium3DTilesetMetadata.prototype, {